回测 - 每日持仓、每日收益率、总收益率

+++ src/main/java/com/bebopze/tdx/quant/dal/entity/BtTaskDO.java
@@ -85,7 +85,7 @@ public class BtTaskDO implements Serializable {
      */
     @TableField("total_day")
     @Schema(description = "总天数")
-    private Short totalDay;
+    private Integer totalDay;
 
     /**
      * 总收益率
